# public/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WAD Judging System API
  version: "1.0.0"
  description: >
    REST API for the WAD Judging System - a role-based platform for managing
    students, events, teams, judges, marks entry, edit-request approvals and
    results for a judged performance competition.
servers:
  - url: /api
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: wad_session
  schemas:
...
    Student:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        fullName: { type: string }
        gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        province:
          type: string
          enum: [WESTERN, CENTRAL, SOUTHERN, NORTHERN, EASTERN, NORTH_WESTERN, NORTH_CENTRAL, UVA, SABARAGAMUWA]
        team: { type: string, enum: [A, B], nullable: true }
        photoUrl: { type: string, nullable: true }
...
security:
  - cookieAuth: []
paths:
...
  /students:
    get:
      tags: [Students]
      summary: List/search students
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Search by name or ID code
        - in: query
          name: gender
          schema: { type: string, enum: [MALE, FEMALE, OTHER] }
        - in: query
          name: province
          schema: { type: string }
        - in: query
          name: team
          schema: { type: string, enum: [A, B] }
      responses:
        "200":
          description: List of students
    post:
      tags: [Students]
      summary: Create a student (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Student" }
      responses:
        "201":
          description: Student created
        "403":
          description: Admin access required
...
  /judges/{id}/performances:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    post:
      tags: [Judges]
      summary: Grant a judge access to a performance (admin only)
      responses:
        "201": { description: Granted }
    delete:
      tags: [Judges]
      summary: Revoke a judge's access to a performance (admin only)
      parameters:
        - in: query
          name: performanceId
          required: true
          schema: { type: string }
      responses:
        "200": { description: Revoked }
...
  /marks:
    get:
      tags: [Marks]
      summary: List mark entries for a performance (judge must be assigned; admin unrestricted)
      parameters:
        - in: query
          name: performanceId
          required: true
          schema: { type: string }
        - in: query
          name: eventId
          schema: { type: string }
        - in: query
          name: studentId
          schema: { type: string }
      responses:
        "200": { description: OK }
        "403": { description: Not assigned to this performance }
    post:
      tags: [Marks]
      summary: Create or update marks (one or more rounds) for a student/event/performance
      description: >
        Creating a brand-new round is always allowed for an assigned judge.
        Editing an existing round requires the judge to hold an APPROVED,
        unconsumed edit request for that mark entry (or be an admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, eventId, performanceId, rounds]
              properties:
                studentId: { type: string }
                eventId: { type: string }
                performanceId: { type: string }
                rounds:
                  type: array
                  items:
                    type: object
                    required: [round, d, e1, e2, e3, e4]
                    properties:
                      round: { type: integer }
                      d: { type: number }
                      dSupervisor: { type: string, nullable: true }
                      e1: { type: number }
                      e2: { type: number }
                      e3: { type: number }
                      e4: { type: number }
                      p: { type: number, default: 0 }
      responses:
        "201": { description: Marks saved }
        "403": { description: Editing requires an approved edit request }
...
  /results:
    get:
      tags: [Results]
      summary: Aggregated results (team totals, top 8, all-rounders) for a performance
      parameters:
        - in: query
          name: performanceId
          required: true
          schema: { type: string }
        - in: query
          name: view
          schema: { type: string, enum: [team, top8, all-rounders] }
        - in: query
          name: eventId
          schema: { type: string }
        - in: query
          name: gender
          schema: { type: string, enum: [MALE, FEMALE, OTHER] }
      responses:
        "200": { description: OK }
  /results/download:
    get:
      tags: [Results]
      summary: Download results as CSV for a performance/gender (optionally one event)
      parameters:
        - in: query
          name: performanceId
          required: true
          schema: { type: string }
        - in: query
          name: gender
          required: true
          schema: { type: string, enum: [MALE, FEMALE, OTHER] }
        - in: query
          name: eventId
          schema: { type: string }
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
...